Day 6 - star 2 slow solution

+++ dotnet/AoC2023/Stars/Day06/Race.cs
@@ -5,6 +5,6 @@ namespace Stars.Day06;
 public class Race
 {
     public int RaceTime { get; set; }
-    public int DistanceRecord { get; set; }
+    public long DistanceRecord { get; set; }
     public int[] WinningOptions { get; set; }
 }
